Updates for EFT Support
Changes question three to ask how they want to receive their
reimbursement. Options are the newly supported EFT and the traditional
printed checks. EFT reimbursements are noted as preferred and included
brief instructions to visit lds.org to setup the reimbursement account
and provides a QR code that can be scanned.